Updated test.py code
Removed commented out code lines to make code
cleaner and easier to read.

Added a variable called 'data' to store the
scraped data into for each scraper.get_data call.
Also, added more print lines to show the type of
variable data is and to format each iteration of
the for loop nicer.